+++ src/Exception/templates/error-page.php
@@ -26,11 +26,6 @@
         .message {
             word-wrap: break-word;
         }
-
-        .msg {
-            color: #FF0000;
-            font-weight: bold;
-        }
     </style>
 </head>