+++ src/ModernPHPException.php
@@ -49,10 +49,37 @@ class ModernPHPException
     protected string $version = "1.0.1";
 
     /**
+     * @var array
+     */
+    private array $config = [
+        "type" => "",
+        "title" => "",
+        "dark_mode" => "",
+        "production_mode" => ""
+    ];
+
+    /**
      * Construct
      */
-    public function __construct()
+    public function __construct(array $config = [])
     {
+        if (!empty($config)) {
+            foreach ($config as $key => $value) {
+                if (!isset($this->config[$key])) {
+                    throw new \Exception("Key '$key' not exists");
+                } else {
+                    $this->config['type'] = $config['type'] ?? "";
+                    $this->config['title'] = $config['title'] ?? "";
+                    $this->config['dark_mode'] = $config['dark_mode'] ?? "";
+                    $this->config['production_mode'] = $config['production_mode'] ?? "";
+                }
+            }
+        }
+
+        if ($this->config['production_mode'] === true) {
+            $this->productionMode();
+        }
+
         $this->bench = new Bench();
         $this->bench->start();
     }