+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Released Notes
 
+## v1.1.0 - (2021-11-01)
+
+## Added
+
+- Added settings directly from the construct
+
+### Fixed
+
+- Fixed CSS color in `info-trace.php`
+
+------------------------------------------------
+
 ## v1.0.1 - (2021-10-29)
 
 ### Fixed